requirements.txt
requests
scrapy
python-dotenv
